refactor(FE): util 폴더 생성
- interface.ts 파일 여기로 옮김
