+++ front/src/util/interface.ts
